requirements.txt
discord.py==1.7.3
sentry-sdk==1.4.3
tortoise-orm==0.17.8
asyncpg==0.24.0
python-dotenv==0.19.1
aerich==0.5.8
aiohttp==3.7.4.post0
cchardet==2.1.7
aiodns==3.0.0
discord-py-interactions==3.0.2
